Refactor Chip component folder structure
