--- src/Chip/common/index.ts
@@ -1,3 +1,0 @@
-export * from "./ChipMovable";
-export * from "./ChipRemovable";
-export * from "./ChipSwatch";
